Move previous draft scenarios to spec/integration
